+++ sensor/mma8451.h
@@ -2,6 +2,6 @@
 #define _MMA8451_H_
 
 void mma8451_init(void);
-void mma8451_get(short *accel);
+void mma8451_get(float *accel);
 
 #endif
